Clean coding - moving all the global config to one place
