+++ auto-chain-agent/agents/src/config/index.ts
@@ -1,8 +1,17 @@
 import dotenv from 'dotenv';
+import path from 'path';
 
 dotenv.config();
 
 export const config = {
+    SUMMARY_FILE_PATH : path.join(process.cwd(), 'summary-differences.json'),
+    CHECK_INTERVAL: 20 * 1000,
+    SUMMARY_DIR: path.join(process.cwd(), 'diffs'),
+    DIFF_FILE_PREFIX: 'summary-diff',
+    LLM_MODEL: "gpt-4-turbo-preview",
+    TEMPERATURE: 0.5,
+    MNEMONIC: process.env.MNEMONIC || '//Alice',
+    NETWORK: 'taurus',
     port: process.env.AGENTS_PORT || 3000,
     anthropicApiKey: process.env.ANTHROPIC_API_KEY,
     openaiApiKey: process.env.OPENAI_API_KEY,
@@ -11,8 +20,4 @@ export const config = {
     autoConsensus: {
         apiKey: process.env.AUTO_CONSENSUS_API_KEY,
     },
-    llmConfig: {
-        temperature: 0.4,
-        maxTokens: 1500
-    }
 };
