+++ auto-chain-agent/agents/src/services/thread/index.ts
@@ -1,0 +1,3 @@
+import { createThreadStorage } from './threadStorage';
+
+export type ThreadStorage = ReturnType<typeof createThreadStorage>;
