+++ auto-chain-agent/agents/src/services/thread/interface.ts
@@ -1,0 +1,28 @@
+import { BaseMessage, MessageContent } from '@langchain/core/messages';
+
+export interface ThreadState {
+    messages: BaseMessage[];
+    toolCalls: Array<{
+        id: string;
+        type: string;
+        function: {
+            name: string;
+            arguments: string;
+        };
+        result?: string;
+    }>;
+}
+
+export interface SummaryDifference {
+    timestamp: string;
+    threadId: string;
+    previousSummary: string | MessageContent;
+    currentSummary: string | MessageContent;
+    difference: string | MessageContent;
+    previousCID?: string;
+}
+
+export interface SummaryState {
+    lastCheck: string;
+    differences: SummaryDifference[];
+}
